WIP simplifying error codes; simplifying data types from u32 to usize; fixing errors

+++ src/beam/compact_term.rs
@@ -91,7 +91,7 @@ pub fn read(r: &mut BinaryReader) -> Hopefully<fterm::FTerm> {
         if index == 0 {
           return Ok(fterm::FTerm::Nil);
         }
-        return Ok(fterm::FTerm::LoadTimeAtom(index as u32))
+        return Ok(fterm::FTerm::LoadTimeAtom(index as usize))
       }
       make_err(CTError::BadAtomTag)
     },
