+++ src/term/fterm.rs
@@ -55,7 +55,7 @@ pub enum FTerm {
   /// A load-time index of label
   LoadTimeLabel(Word),
   /// A load-time atom index in the loader atom table
-  LoadTimeAtom(u32),
+  LoadTimeAtom(usize),
   // /// A load-time word value literally specified
   //LoadTimeInt(SWord),
   /// A load-time index in literal heap
